Updated the gitignore, added time slow.
The gitignore was in the wrong place and I don';t have time to fix the added files that don't need to be in the repo right now.
Time slow functionality works, there's duration to it and you can control how slow things go -ish. There's also a cooldown. I'll need to implement the weapon later, but the good news is that time can be slowed for any actor(s).