Shooting Functionality implemented. Spawner and BouncyBalls implemented as test objects.
An orb grows in size up to a certain point in front of the character while the player is holding the shoot button. Once the player releases the button, the orb moves at a speed that depends on how long the shoot button was held down up to a certain point away from the player. When the orb hits something, it detects if the player if in the sphere of influence. If the player will be affected, time slows for everything except the player. If the player is not in the sphere of influence, time slows for everything inside of the sphere of influence. The time slow duration, magnitude, and radius of the sphere of influence are all dependent on how long the player held down the shoot button.
If the orb continues for some time without hitting anything, it dissipates. (the object is destroyed.)