Menus, Jumping, Music, Winning, Losing, Pausing, And Save Game Data.
Main Menu can start the game, select a level, quit the game, and has a lovely view of level 1 in the background.
The player oofs when they jump. May change that sound effect.
There are three music tracks randomly chosen at the start of every section, including Main Menu. You cannot choose which one that will be. Pausing pauses the music. Playing plays the music.
You can lose now! When the player dies (via the health system, not sure if there are any ways to fall into the abyss otherwise... They will get the option to return to the main menu or restart the level.
You can win, if you want. You get the option to go to the main menu, restart the level, on soldier mightily on to the test dummy level.
The pause menu pauses the game, the music, and allows you to choose from restarting the level, going to the main menu, or resuming play.
The only way to unlock level 2 is by winning level 1. Once you've won one, you can play two directly from the level select in the main menu.