BaseEnemy, AISpawner, Test Level.
Everything builds! The AI in the Contents/Blueprint/AI folder can be placed either solo by dragging in the character, or in increasingly large groups by dragging in the spawner. BOTH OPTIONS NEED A NAV MESH BOUNDS VOLUME TO WORK. And make sure to orient them in the direction where they will see the player, or the player will literally have to bump into them in order for anything to happen.
They can kill you quickly in large groups. Default damage is -10, so ten hits and you're toast.