Collapsing platform component, builds for the entire game, updates to the timed spawner and thirdpersoncharacter.
Collapsing platforms of all types incorporate the componentcollapse, which also serves as a moving platform component until told to collapse. On trigger of overlap, the component counts down until it collapses. On trigger end overlap, this countdown resets. When the actor is collapsing, visual and auditory feedback occurs, when the actor collapses, visual and auditory feedback occurs as well, and then gravity sets in, the floor falls, and after a delay the actor and all components are destroyed.
Timedspawner and thirdpersoncharacter are more designer-friendly with regards to letting variables be adjustable from the level view.