Winning can unlock multiple paths, Main Menu has a secret. Lvl 1 lighting does turn off unexpectedly.
This was a last-minute fix to incorporate the level test into the design without messing with too much of the game. So the first level in the (new!) array of levels that get unlocked is opened upon clicking continue. All of the other levels are unlocked in the main menu level select. So, for example, Level 3 "unlocks" and goes to level 1, because level 4 is incomplete and isn't ready yet, but level 4 also unlocks level 4, so players can go to it if they want.
Level 3 also unlocks the test level, hidden at the bottom of the scroll menu in the level select.