Added itch.io to Readme, made some fixes to MainMenu, tried to fix AI in LVL1, LVL2 needed to be rebuilt.
The link to Itch.io in the readme is there so that people who are interested in the game but not interested in deving in unreal can play a version of it as well, if not  the most up-to-date version of it.
MainMenu needed to only open the correct level select, it was openjing both the griod layout and the scrollbar.
AI are still not fixed in LVL 1. I tried reconfiguring the AI, placing them solidly on the ground, rebuilding the navmesh, and repackaging in multiple ways as well as trying it through the project launcher.
Rebuilt LVL2.